attendance_recognition/requirements.txt
streamlit==1.39.0
SQLAlchemy==2.0.36
numpy==1.24.2
pillow==10.2.0
pickleshare==0.7.5
sniffio==1.3.1*
face-recognition==1.3.0
pytz==2023.2
regex==2024.9.11
pandas==1.5.3
fpdf==1.7.2
PyMySQL==1.1.1

